extract Express app
