+++ ressaite-api/src/index.ts
@@ -1,6 +1,6 @@
-import express from "express";
 import { sequelize } from "./core/db/instance";
 import { faireUnPouet } from "@al-un/ressaite-core/pouet";
+import app from "./app";
 
 (async () => {
   try {
@@ -11,28 +11,7 @@ import { faireUnPouet } from "@al-un/ressaite-core/pouet";
   }
 })();
 
-const app = express();
-
-// import CookieParser from "cookie-parser";
-// import BodyParser from "body-parser";
-// app.use(CookieParser());
-// app.use(BodyParser.urlencoded({ extended: true }));
-app.use(express.json());
-
 const port = 8000;
-
-import AuthRouter from "@/um/routers/AuthRouter";
-import AuthMiddleware from "@/um/middleware/AuthMiddleware";
-
-app.get("/helloworld", (req, res) => {
-  res.send("helloworld\n");
-});
-
-app.use(AuthRouter);
-
-app.get("/pouet", AuthMiddleware.authenticate, function (req, res) {
-  res.send("YAYY\n");
-});
 
 app.listen(port, () => {
   faireUnPouet();
